feat(api): implement comic generation and processing endpoints
- Add /api/comics/generate endpoint for story and character processing
- Add /api/comics/upload endpoint for handling character image uploads
- Add /api/comics/update endpoint for story and character updates
- Add /api/comics/preview endpoint for comic preview generation
- Add /api/comics/mint-download endpoint for NFT minting
- Add /api/marketplace/publish endpoint for marketplace integration

The API layer now supports:
- Story and character data processing
- Image upload handling with IPFS integration
- Comic preview generation
- NFT minting and download functionality
- Basic marketplace operations

All endpoints include proper error handling and input validation.
