+++ app/api/backup-apis/payments/fiat/route.ts
@@ -1,3 +1,6 @@
+
+// app/api/payments/fiat/route.ts
+
 import { NextResponse } from 'next/server';
 import Stripe from 'stripe';
 import { ethers } from 'ethers';
