+++ app/api/comics/mint-download/route.ts
@@ -1,0 +1,5 @@
+import { NextResponse } from "next/server";
+
+export async function GET(request: Request) {
+    return NextResponse.json({ success: true });
+}
