+++ app/api/comics/preview/route.ts
@@ -1,33 +1,13 @@
-import { NextResponse } from 'next/server';
-import { ComicMetadata } from '@/types/comic';
+import { NextResponse } from "next/server";
 
-export async function POST(request: Request) {
-    try {
-        const body = await request.json();
-        const { comic } = body as { comic: ComicMetadata };
-
-        // Validate comic metadata
-        if (!comic || !comic.name || !comic.description || !comic.image) {
-            return NextResponse.json({ error: 'Invalid comic metadata' }, { status: 400 });
-        }
-
-        // Return preview (in production, this could render a temporary view)
-        return NextResponse.json({
-            success: true,
-            preview: {
-                ...comic,
-                pages: comic.description.split('\n').map((text, i) => ({
-                    pageNumber: i + 1,
-                    text,
-                    image: comic.image,
-                })),
-            },
-        });
-    } catch (error: any) {
-        console.error('Comic preview error:', error);
-        return NextResponse.json(
-            { error: `Failed to generate preview: ${error.message || 'Unknown error'}` },
-            { status: 500 }
-        );
-    }
+export async function GET() {
+    return NextResponse.json({
+        pages: [
+            { image: "/mock-image1.jpg", text: "Rodgers cooks a delicious meal." },
+            { image: "/mock-image2.jpg", text: "Playing badminton with friends." },
+            { image: "/mock-image3.jpg", text: "A sunny day at the park." },
+            { image: "/mock-image4.jpg", text: "Exploring the forest." },
+            { image: "/mock-image5.jpg", text: "Celebrating a birthday." },
+        ],
+    });
 }
