+++ app/api/comics/update/route.ts
