+++ app/api/marketplace/publish/route.ts
@@ -1,0 +1,5 @@
+import { NextResponse } from "next/server";
+
+export async function POST(request: Request) {
+    return NextResponse.json({ success: true });
+}
